requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
google-generativeai
beautifulsoup4